PlayServices check before activation
