+++ app/src/main/res/navigation/nav_graph.xml
@@ -44,6 +44,13 @@
         android:label="@string/app_name"
         tools:layout="@layout/fragment_welcome">
         <action
+            android:id="@+id/action_nav_welcome_to_nav_play_services_update"
+            app:destination="@id/nav_play_services_update"
+            app:enterAnim="@anim/nav_default_enter_anim"
+            app:exitAnim="@anim/nav_default_exit_anim"
+            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
+        <action
             android:id="@+id/action_nav_welcome_fragment_to_nav_legacy_update"
             app:destination="@+id/nav_legacy_update_fragment"
             app:enterAnim="@anim/nav_default_enter_anim"
@@ -170,6 +177,13 @@
         android:label="@string/app_name"
         tools:layout="@layout/fragment_dashboard">
         <action
+            android:id="@+id/action_nav_dashboard_to_nav_play_services_update"
+            app:destination="@id/nav_play_services_update"
+            app:enterAnim="@anim/nav_default_enter_anim"
+            app:exitAnim="@anim/nav_default_exit_anim"
+            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
+        <action
             android:id="@+id/action_nav_dashboard_to_nav_send_data"
             app:destination="@+id/nav_send_data"
             app:enterAnim="@anim/nav_default_enter_anim"
@@ -209,13 +223,6 @@
         <action
             android:id="@+id/action_nav_dashboard_to_nav_exposures"
             app:destination="@id/nav_exposures"
-            app:enterAnim="@anim/nav_default_enter_anim"
-            app:exitAnim="@anim/nav_default_exit_anim"
-            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
-            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
-        <action
-            android:id="@+id/action_nav_dashboard_to_nav_play_services_update"
-            app:destination="@id/nav_play_services_update"
             app:enterAnim="@anim/nav_default_enter_anim"
             app:exitAnim="@anim/nav_default_exit_anim"
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
